Add files via upload
In this change, new additions are modification of "StrengthChecker" class and 4 necessary metrics to assist it ("ShannonEntropy" for randomness checking, "KeyboardPatternDetect" to detect common sequential patterns, "RepetitiveCharacterDetect" to identify more than two consecutive characters and "LengthDetect" to check the length). Masking has been applied, and relevant information can be visible in the console after user presses a certain key.